// src/routes/user/user-types.ts
export enum GetMemoryCountError {
  USER_NOT_FOUND = "USER_NOT_FOUND",
}

export enum UpdateUserProfileError {
  USER_NOT_FOUND = "USER_NOT_FOUND",
  INVALID_EMAIL = "INVALID_EMAIL",
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  INVALID_PASSWORD = "INVALID_PASSWORD",
  PASSWORD_TOO_WEAK = "PASSWORD_TOO_WEAK",
  UNAUTHORIZED = "UNAUTHORIZED",
  INVALID_NAME = "INVALID_NAME",
  SAME_AS_PREVIOUS_NAME = "SAME_AS_PREVIOUS_NAME",
  UNKNOWN = "UNKNOWN",
}


export enum DeleteUserError {
  USER_NOT_FOUND = "USER_NOT_FOUND",
  USER_HAS_MEMORIES = "USER_HAS_MEMORIES",
  USER_HAS_CHATS = "USER_HAS_CHATS",
  UNKNOWN = "UNKNOWN",
}